+++ ebean-test/src/test/java/org/tests/query/aggregation/TestAggregationCount.java
@@ -13,6 +13,7 @@ import org.tests.model.basic.ResetBasicData;
 import org.tests.model.tevent.TEventMany;
 import org.tests.model.tevent.TEventOne;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.List;
@@ -24,15 +25,15 @@ public class TestAggregationCount extends BaseTestCase {
   @BeforeAll
   public static void setup() {
     TEventOne one = new TEventOne("first", TEventOne.Status.AA);
-    one.getLogs().add(new TEventMany("all", 1, 10));
-    one.getLogs().add(new TEventMany("be", 2, 12.2));
-    one.getLogs().add(new TEventMany("add", 3, 13));
+    one.getLogs().add(new TEventMany("all", 1, BigDecimal.valueOf(10)));
+    one.getLogs().add(new TEventMany("be", 2, BigDecimal.valueOf(12.2)));
+    one.getLogs().add(new TEventMany("add", 3, BigDecimal.valueOf(13)));
     DB.save(one);
 
     TEventOne two = new TEventOne("second", TEventOne.Status.AA);
-    two.getLogs().add(new TEventMany("at", 10, 10));
-    two.getLogs().add(new TEventMany("add", 30, 13));
-    two.getLogs().add(new TEventMany("alf", 30, 13));
+    two.getLogs().add(new TEventMany("at", 10, BigDecimal.valueOf(10)));
+    two.getLogs().add(new TEventMany("add", 30, BigDecimal.valueOf(13)));
+    two.getLogs().add(new TEventMany("alf", 30, BigDecimal.valueOf(13)));
     DB.save(two);
 
     TEventOne three = new TEventOne("thrird", TEventOne.Status.BB);
@@ -53,7 +54,7 @@ public class TestAggregationCount extends BaseTestCase {
     for (TEventOne eventOne : list) {
       // lazy loading on Aggregation properties
       // is not expected to work at this stage
-      Double totalAmount = eventOne.getTotalAmount();
+      BigDecimal totalAmount = eventOne.getTotalAmount();
       assertThat(totalAmount).isNull();
     }
   }
